# gen/http/openapi3.json
{
  "openapi": "3.0.3",
  "info": {
    "title": "Fire Tracker Service",
    "description": "Service for fire tracking",
    "termsOfService": "terms",
    "contact": {
      "name": "support",
      "url": "firetracker.com",
      "email": "support@firetracker.com"
    },
    "license": {
      "name": "GPT-3",
      "url": "licence url"
    },
    "version": "1.0"
  },
  "servers": [
    {
      "url": "http://localhost:8080"
    },
    {
      "url": "https://test.tmf.com"
    },
    {
      "url": "https://trackmyfire.com"
    }
  ],
  "paths": {
...
    "/weather": {
      "post": {
        "tags": [
          "weather"
        ],
        "summary": "create weather",
        "description": "Create a weather observation and optional payloads",
        "operationId": "weather#create",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/Weather"
              },
              "example": {
                "createdAt": "2020-01-02T00:00:00Z",
                "dewPoint": {
                  "unit": "K",
                  "value": 20
                },
                "fireID": 1,
                "humidity": 1500,
                "id": 1,
                "logID": 1,
                "temperature": {
                  "unit": "K",
                  "value": 20
                },
                "weatherTime": "2020-01-02T00:00:00Z",
                "weatherType": "Sunny",
                "wind": {
                  "direction": "SE",
                  "speed": 15,
                  "unit": "KPH"
                }
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "OK response.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Weather"
                },
                "example": {
                  "createdAt": "2020-01-02T00:00:00Z",
                  "dewPoint": {
                    "unit": "C",
                    "value": 20
                  },
                  "fireID": 1,
                  "humidity": 1500,
                  "id": 1,
                  "logID": 1,
                  "temperature": {
                    "unit": "C",
                    "value": 20
                  },
                  "weatherTime": "2020-01-02T00:00:00Z",
                  "weatherType": "Cloudy",
                  "wind": {
                    "direction": "E",
                    "speed": 15,
                    "unit": "MPH"
                  }
                }
              }
            }
          },
          "400": {
            "description": "Bad Request response.",
            "content": {
              "application/vnd.goa.error": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                },
                "example": {
                  "id": "3F1FKVRR",
                  "message": "Value of ID must be an integer",
                  "name": "bad_request"
                }
              }
            }
          },
          "500": {
            "description": "Internal Server Error response.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Weather"
                },
                "example": {
                  "createdAt": "2020-01-02T00:00:00Z",
                  "dewPoint": {
                    "unit": "C",
                    "value": 20
                  },
                  "fireID": 1,
                  "humidity": 1500,
                  "id": 1,
                  "logID": 1,
                  "temperature": {
                    "unit": "C",
                    "value": 20
                  },
                  "weatherTime": "2020-01-02T00:00:00Z",
                  "weatherType": "Raining",
                  "wind": {
                    "direction": "E",
                    "speed": 15,
                    "unit": "MPH"
                  }
                }
              }
            }
          }
        }
      }
    },
    "/weather/list": {
      "post": {
        "tags": [
          "weather"
        ],
        "summary": "list weather",
        "description": "List weather",
        "operationId": "weather#list",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ListRequestBody2"
              },
              "example": {
                "filters": {
                  "humidity": [
                    {
                      "key": "id",
                      "operator": "\u003c",
                      "value": 1
                    },
                    {
                      "key": "id",
                      "operator": "\u003c",
                      "value": 1
                    },
                    {
                      "key": "id",
                      "operator": "\u003c",
                      "value": 1
                    },
                    {
                      "key": "id",
                      "operator": "\u003c",
                      "value": 1
                    }
                  ],
                  "temperature": [
                    {
                      "key": "id",
                      "operator": "\u003c",
                      "value": 1
                    },
                    {
                      "key": "id",
                      "operator": "\u003c",
                      "value": 1
                    }
                  ],
                  "time": [
                    {
                      "key": "added_at",
                      "operator": "=",
                      "value": "2020-01-01T00:00:00Z"
                    },
                    {
                      "key": "added_at",
                      "operator": "=",
                      "value": "2020-01-01T00:00:00Z"
                    },
                    {
                      "key": "added_at",
                      "operator": "=",
                      "value": "2020-01-01T00:00:00Z"
                    }
                  ],
                  "windSpeed": [
                    {
                      "key": "id",
                      "operator": "\u003c",
                      "value": 1
                    },
                    {
                      "key": "id",
                      "operator": "\u003c",
                      "value": 1
                    }
                  ]
                },
                "pagination": {
                  "limit": 10,
                  "page": 1
                },
                "search": {
                  "description": "Almost blew the fire out!",
                  "name": "Windy Part"
                },
                "sort": {
                  "id": "ASC, DESC"
                }
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "OK response.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/WeatherList"
                },
                "example": {
                  "pagination": {
                    "limit": 10,
                    "page": 1,
                    "total": 15
                  },
                  "weathers": [
                    {
                      "createdAt": "2020-01-02T00:00:00Z",
                      "dewPoint": {
                        "unit": "C",
                        "value": 20
                      },
                      "fireID": 1,
                      "humidity": 1500,
                      "id": 1,
                      "logID": 1,
                      "temperature": {
                        "unit": "C",
                        "value": 20
                      },
                      "weatherTime": "2020-01-02T00:00:00Z",
                      "weatherType": "Sunny",
                      "wind": {
                        "direction": "E",
                        "speed": 15,
                        "unit": "MPH"
                      }
                    },
                    {
                      "createdAt": "2020-01-02T00:00:00Z",
                      "dewPoint": {
                        "unit": "C",
                        "value": 20
                      },
                      "fireID": 1,
                      "humidity": 1500,
                      "id": 1,
                      "logID": 1,
                      "temperature": {
                        "unit": "C",
                        "value": 20
                      },
                      "weatherTime": "2020-01-02T00:00:00Z",
                      "weatherType": "Sunny",
                      "wind": {
                        "direction": "E",
                        "speed": 15,
                        "unit": "MPH"
                      }
                    }
                  ]
                }
              }
            }
          },
          "400": {
            "description": "Bad Request response.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/WeatherList"
                },
                "example": {
                  "pagination": {
                    "limit": 10,
                    "page": 1,
                    "total": 15
                  },
                  "weathers": [
                    {
                      "createdAt": "2020-01-02T00:00:00Z",
                      "dewPoint": {
                        "unit": "C",
                        "value": 20
                      },
                      "fireID": 1,
                      "humidity": 1500,
                      "id": 1,
                      "logID": 1,
                      "temperature": {
                        "unit": "C",
                        "value": 20
                      },
                      "weatherTime": "2020-01-02T00:00:00Z",
                      "weatherType": "Sunny",
                      "wind": {
                        "direction": "E",
                        "speed": 15,
                        "unit": "MPH"
                      }
                    },
                    {
                      "createdAt": "2020-01-02T00:00:00Z",
                      "dewPoint": {
                        "unit": "C",
                        "value": 20
                      },
                      "fireID": 1,
                      "humidity": 1500,
                      "id": 1,
                      "logID": 1,
                      "temperature": {
                        "unit": "C",
                        "value": 20
                      },
                      "weatherTime": "2020-01-02T00:00:00Z",
                      "weatherType": "Sunny",
                      "wind": {
                        "direction": "E",
                        "speed": 15,
                        "unit": "MPH"
                      }
                    },
                    {
                      "createdAt": "2020-01-02T00:00:00Z",
                      "dewPoint": {
                        "unit": "C",
                        "value": 20
                      },
                      "fireID": 1,
                      "humidity": 1500,
                      "id": 1,
                      "logID": 1,
                      "temperature": {
                        "unit": "C",
                        "value": 20
                      },
                      "weatherTime": "2020-01-02T00:00:00Z",
                      "weatherType": "Sunny",
                      "wind": {
                        "direction": "E",
                        "speed": 15,
                        "unit": "MPH"
                      }
                    },
                    {
                      "createdAt": "2020-01-02T00:00:00Z",
                      "dewPoint": {
                        "unit": "C",
                        "value": 20
                      },
                      "fireID": 1,
                      "humidity": 1500,
                      "id": 1,
                      "logID": 1,
                      "temperature": {
                        "unit": "C",
                        "value": 20
                      },
                      "weatherTime": "2020-01-02T00:00:00Z",
                      "weatherType": "Sunny",
                      "wind": {
                        "direction": "E",
                        "speed": 15,
                        "unit": "MPH"
                      }
                    }
                  ]
                }
              }
            }
          },
          "500": {
            "description": "Internal Server Error response.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/WeatherList"
                },
                "example": {
                  "pagination": {
                    "limit": 10,
                    "page": 1,
                    "total": 15
                  },
                  "weathers": [
                    {
                      "createdAt": "2020-01-02T00:00:00Z",
                      "dewPoint": {
                        "unit": "C",
                        "value": 20
                      },
                      "fireID": 1,
                      "humidity": 1500,
                      "id": 1,
                      "logID": 1,
                      "temperature": {
                        "unit": "C",
                        "value": 20
                      },
                      "weatherTime": "2020-01-02T00:00:00Z",
                      "weatherType": "Sunny",
                      "wind": {
                        "direction": "E",
                        "speed": 15,
                        "unit": "MPH"
                      }
                    },
                    {
                      "createdAt": "2020-01-02T00:00:00Z",
                      "dewPoint": {
                        "unit": "C",
                        "value": 20
                      },
                      "fireID": 1,
                      "humidity": 1500,
                      "id": 1,
                      "logID": 1,
                      "temperature": {
                        "unit": "C",
                        "value": 20
                      },
                      "weatherTime": "2020-01-02T00:00:00Z",
                      "weatherType": "Sunny",
                      "wind": {
                        "direction": "E",
                        "speed": 15,
                        "unit": "MPH"
                      }
                    },
                    {
                      "createdAt": "2020-01-02T00:00:00Z",
                      "dewPoint": {
                        "unit": "C",
                        "value": 20
                      },
                      "fireID": 1,
                      "humidity": 1500,
                      "id": 1,
                      "logID": 1,
                      "temperature": {
                        "unit": "C",
                        "value": 20
                      },
                      "weatherTime": "2020-01-02T00:00:00Z",
                      "weatherType": "Sunny",
                      "wind": {
                        "direction": "E",
                        "speed": 15,
                        "unit": "MPH"
                      }
                    }
                  ]
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "Error": {
        "type": "object",
        "properties": {
          "fault": {
            "type": "boolean",
            "description": "Is the error a server-side fault?",
            "example": false
          },
          "id": {
            "type": "string",
            "description": "ID is a unique identifier for this particular occurrence of the problem.",
            "example": "123abc"
          },
          "message": {
            "type": "string",
            "description": "Message is a human-readable explanation specific to this occurrence of the problem.",
            "example": "parameter 'p' must be an integer"
          },
          "name": {
            "type": "string",
            "description": "Name is the name of this class of errors.",
            "example": "bad_request"
          },
          "temporary": {
            "type": "boolean",
            "description": "Is the error temporary?",
            "example": true
          },
          "timeout": {
            "type": "boolean",
            "description": "Is the error a timeout?",
            "example": true
          }
        },
        "description": "Please ensure you are sending valid data.",
        "example": {
          "id": "3F1FKVRR",
          "message": "Value of ID must be an integer",
          "name": "bad_request"
        },
        "required": [
          "name",
          "id",
          "message",
          "temporary",
          "timeout",
          "fault"
        ]
      },
...
      "ListRequestBody2": {
        "type": "object",
        "properties": {
          "filters": {
            "$ref": "#/components/schemas/WeatherFilters"
          },
          "pagination": {
            "$ref": "#/components/schemas/LogPagination"
          },
          "search": {
            "$ref": "#/components/schemas/LogSearch"
          },
          "sort": {
            "$ref": "#/components/schemas/WeatherSorts"
          }
        },
        "example": {
          "filters": {
            "humidity": [
              {
                "key": "id",
                "operator": "\u003c",
                "value": 1
              },
              {
                "key": "id",
                "operator": "\u003c",
                "value": 1
              },
              {
                "key": "id",
                "operator": "\u003c",
                "value": 1
              },
              {
                "key": "id",
                "operator": "\u003c",
                "value": 1
              }
            ],
            "temperature": [
              {
                "key": "id",
                "operator": "\u003c",
                "value": 1
              },
              {
                "key": "id",
                "operator": "\u003c",
                "value": 1
              }
            ],
            "time": [
              {
                "key": "added_at",
                "operator": "=",
                "value": "2020-01-01T00:00:00Z"
              },
              {
                "key": "added_at",
                "operator": "=",
                "value": "2020-01-01T00:00:00Z"
              },
              {
                "key": "added_at",
                "operator": "=",
                "value": "2020-01-01T00:00:00Z"
              }
            ],
            "windSpeed": [
              {
                "key": "id",
                "operator": "\u003c",
                "value": 1
              },
              {
                "key": "id",
                "operator": "\u003c",
                "value": 1
              }
            ]
          },
          "pagination": {
            "limit": 10,
            "page": 1
          },
          "search": {
            "description": "Almost blew the fire out!",
            "name": "Windy Part"
          },
          "sort": {
            "id": "ASC, DESC"
          }
        },
        "required": [
          "filters",
          "search",
          "sort",
          "pagination"
        ]
      },
...
      "Weather": {
        "type": "object",
        "properties": {
          "createdAt": {
            "type": "string",
            "description": "name",
            "example": "2020-01-02T00:00:00Z",
            "format": "date-time"
          },
          "dewPoint": {
            "$ref": "#/components/schemas/Temperature"
          },
          "fireID": {
            "type": "integer",
            "description": "ID of the fire",
            "example": 1,
            "format": "int64"
          },
          "humidity": {
            "type": "integer",
            "description": "humidity level",
            "example": 1500,
            "format": "int32"
          },
          "id": {
            "type": "integer",
            "description": "id",
            "example": 1,
            "format": "int64"
          },
          "logID": {
            "type": "integer",
            "description": "ID of the log",
            "example": 1,
            "format": "int64"
          },
          "temperature": {
            "$ref": "#/components/schemas/Temperature"
          },
          "weatherTime": {
            "type": "string",
            "example": "2020-01-02T00:00:00Z",
            "format": "date-time"
          },
          "weatherType": {
            "type": "string",
            "description": "type of weather",
            "example": "Sunny",
            "enum": [
              "Sunny",
              "Cloudy",
              "Raining",
              "Windy"
            ]
          },
          "wind": {
            "$ref": "#/components/schemas/Wind"
          }
        },
        "example": {
          "createdAt": "2020-01-02T00:00:00Z",
          "dewPoint": {
            "unit": "K",
            "value": 20
          },
          "fireID": 1,
          "humidity": 1500,
          "id": 1,
          "logID": 1,
          "temperature": {
            "unit": "K",
            "value": 20
          },
          "weatherTime": "2020-01-02T00:00:00Z",
          "weatherType": "Windy",
          "wind": {
            "direction": "SE",
            "speed": 15,
            "unit": "KPH"
          }
        }
      },
...
    }
  }
}